+++ lib/src/models/recurrent/recurrent_request.dart
@@ -21,7 +21,7 @@ class RecurrentRequest with _$RecurrentRequest {
     required String description,
 
     /// Сумма платежа. Должна быть больше 0
-    required int amount,
+    required String amount,
 
     /// Дата и время первого платежа по плану во временной зоне UTC.
     ///