+++ CHANGELOG.md
@@ -5,3 +5,11 @@
 ## 1.0.1
 
 - Fix README.md and chore
+
+## 1.0.2
+
+- Change amount type to String
+
+## 1.0.3
+
+- Change reccurrent logic